add: vue components and styles

+++ src/backend/database/migrations/2019_01_12_191142_create_tasks_table.php
@@ -14,8 +14,8 @@ class CreateTasksTable extends Migration
     public function up()
     {
         Schema::create('tasks', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('project_id');
+            $table->increments('id');
+            $table->unsignedInteger('project_id');
             $table->text('body');
             $table->boolean('completed')->default(false);
             $table->timestamps();
